Move R related functions to R.py
